# backend/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "このswaggerはCampusCrowdMonitorのapiです",
        "title": "CampusCrowdMonitor",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/staycount/history/{room_id}": {
            "get": {
                "description": "滞在者数の履歴を取得する。これは24時間分のデータを取得する。",
                "produces": [
                    "application/json"
                ],
                "summary": "滞在者数の履歴を取得する。部屋指定かつ日付指定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "日付データを記載する 例: 2021-10-01",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetStayCountHistoryModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.GetStayCountHistoryModel": {
            "type": "object",
            "properties": {
                "0": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}